+++ Gp-3/Gp-3.csproj
@@ -5,9 +5,11 @@
     <UserSecretsId>aspnet-Gp_3-88BC277A-695A-414A-A56F-5F007DA456C1</UserSecretsId>
     <RootNamespace>Gp_3</RootNamespace>
     <OutputType>Exe</OutputType>
+    <StartupObject></StartupObject>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="bootstrap" Version="4.6.0" />
     <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="5.0.4" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="5.0.4" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="5.0.4" />
