--- Gp-3/Output-Build.txt
@@ -1,1 +1,0 @@
-﻿
