+++ Gp-3/Views/Account/Login.cshtml
@@ -3,11 +3,12 @@
     ViewBag.Title = "Login";
 }
 
-<h1>User Login</h1>
+
 
 <div class="row">
-    <div class="col-md-12">
-        <form method="post">
+    <div class="col-md-12" align="center"style="margin-top:100px">
+        <h1>User Login</h1>
+        <form method="post" style="width:500px">
             <div asp-validation-summary="All" class="text-danger"></div>
             <div class="form-group">
                 <label asp-for="Email"></label>
@@ -28,7 +29,11 @@
                 </div>
             </div>
             <button type="submit" class="btn btn-primary">Login</button>
+            <button type="button" class="btn btn-primary">Register</button><br />
+            <a href="">forget password?</a>
         </form>
+        
+
     </div>
 </div>
